staking.tsx breakdown into 4 components
